Refactored Virtual File System into Modular Structure
